Refactor: Moved login component from components directory to features directory
