+++ src/features/Login/components/SocialAuth.jsx
@@ -1,7 +1,7 @@
 import { SocialAuthContainer, ButtonContainer } from './socialauth.styled';
 import LoginHeader from './LoginHeader';
 import SocialAuthButton from './SocialAuthButton';
-import authProviders from '../../../utils/authProviders';
+import authProviders from '../utils/authProviders';
 
 const LoginForm = () => (
   <SocialAuthContainer>
